Refactor parsing to subcommand classes
