+++ src/main/java/org/dita/dost/invoker/DeliverablesArguments.java
@@ -8,114 +8,54 @@
 
 package org.dita.dost.invoker;
 
+import org.apache.tools.ant.BuildException;
+
 import java.io.File;
-import java.util.List;
+import java.util.ArrayDeque;
+import java.util.Arrays;
+import java.util.Deque;
 import java.util.Map;
-import java.util.Vector;
 
 public class DeliverablesArguments extends Arguments {
 
-    /**
-     * Project file
-     */
-    final File projectFile;
+    File projectFile;
 
-    public DeliverablesArguments(File projectFile,
-                                 boolean useColor, int msgOutputLevel, File buildFile, Vector<String> targets,
-                                 Vector<String> listeners, Vector<String> propertyFiles, boolean allowInput, boolean keepGoingMode,
-                                 String loggerClassname, String inputHandlerClassname, boolean emacsMode, Integer threadPriority,
-                                 boolean proxy, boolean justPrintUsage, boolean justPrintVersion, boolean justPrintDiagnostics,
-                                 File logFile, Map<String, Object> definedProps) {
-        super(useColor, msgOutputLevel, buildFile, targets, listeners, propertyFiles, allowInput, keepGoingMode,
-                loggerClassname, inputHandlerClassname, emacsMode, threadPriority, proxy, justPrintUsage, justPrintVersion,
-                justPrintDiagnostics, logFile, definedProps);
-        this.projectFile = projectFile;
+    @Override
+    DeliverablesArguments parse(final String[] arguments) {
+        useColor = getUseColor();
+        final Deque<String> args = new ArrayDeque<>(Arrays.asList(arguments));
+        while (!args.isEmpty()) {
+            final String arg = args.pop();
+            if (arg.equals("deliverables") || isLongForm(arg, "-deliverables")) {
+                // ignore
+            } else if (isLongForm(arg, "-project") || arg.equals("-p")) {
+                handleArgProject(arg, args);
+            } else {
+                parseCommonOptions(arg, args);
+            }
+        }
+        return this;
+    }
+
+    /**
+     * Handle the --project argument
+     */
+    private void handleArgProject(final String arg, final Deque<String> args) {
+        final Map.Entry<String, String> entry = parse(arg, args);
+        if (entry.getValue() == null) {
+            throw new BuildException("Missing value for project " + entry.getKey());
+        }
+        projectFile = new File(entry.getValue()).getAbsoluteFile();
     }
 
     @Override
     void printUsage() {
         final StringBuilder msg = new StringBuilder();
         msg.append("Usage: dita deliverables [options]\n");
-//        msg.append("   or: dita --project=<file> [options]\n");
-//        msg.append("   or: dita --propertyfile=<file> [options]\n");
-//        msg.append("   or: dita --install [=<file> | <url> | <id>]\n");
-//        msg.append("   or: dita --uninstall <id>\n");
-//        msg.append("   or: dita --plugins\n");
-//        msg.append("   or: dita --transtypes\n");
-//        msg.append("   or: dita --deliverables\n");
-//        msg.append("   or: dita --help\n");
-//        msg.append("   or: dita --version\n");
-//        msg.append("Subcommands: \n");
-////        msg.append("   or: dita --project=<file> [options]\n");
-////        msg.append("   or: dita --propertyfile=<file> [options]\n");
-//        msg.append("   install [=<file> | <url> | <id>]\n");
-//        msg.append("   uninstall <id>\n");
-//        msg.append("   plugins\n");
-//        msg.append("   transtypes\n");
-//        msg.append("   deliverables\n");
-////        msg.append("   or: dita --help\n");
-////        msg.append("   or: dita --version\n");
-//        msg.append("Arguments: \n");
-//        msg.append("  -i <file>, --input=<file>    input file\n");
-//        msg.append("  -f <name>, --format=<name>   output format (transformation type)\n");
-//        msg.append("  -p <name>, --project=<name>  run project file\n");
-//        msg.append("  -r <file>, --resource=<file> resource file\n");
-////        msg.append("  --install [<file>]           install plug-in from a local ZIP file\n");
-////        msg.append("  --install [<url>]            install plug-in from a URL\n");
-////        msg.append("  --install [<id>]             install plug-in from plugin registry\n");
-////        msg.append("  --install                    reload plug-ins\n");
-////        msg.append("  --uninstall <id>             uninstall plug-in with the ID\n");
-////        msg.append("  --plugins                    print list of installed plug-ins\n");
-////        msg.append("  --transtypes                 print list of installed transtypes\n");
-////        msg.append("  --deliverables               print list of deliverables in project\n");
-//        msg.append("  -h, --help                   print this message\n");
-//        msg.append("  --version                    print version information and exit\n");
         msg.append("Options: \n");
-//        msg.append("  -o, --output=<dir>          output directory\n");
-        // msg.append("  -diagnostics           print information that might be helpful to"
-        // + lSep);
-        // msg.append("                         diagnose or report problems." +
-        // lSep);
-        // msg.append("  -quiet, -q             be extra quiet" + lSep);
-//        msg.append("  --filter=<files>             filter and flagging files\n");
-//        msg.append("  --force                      force install plug-in\n");
-//        msg.append("  -t, --temp=<dir>             temporary directory\n");
-        msg.append("  -v, --verbose                verbose logging\n");
         msg.append("  -d, --debug                  print debugging information\n");
-        // msg.append("  -emacs, -e             produce logging information without adornments"
-        // + lSep);
-        // msg.append("  -lib <path>            specifies a path to search for jars and classes"
-        // + lSep);
-//        msg.append("  -l, --logfile=<file>        use given file for log\n");
-        // msg.append("  -logger <classname>    the class which is to perform logging"
-        // + lSep);
-        // msg.append("  -listener <classname>  add an instance of class as a project listener"
-        // + lSep);
-        // msg.append("  -noinput               do not allow interactive input"
-        // + lSep);
-        // msg.append("  -buildfile <file>      use given buildfile" + lSep);
-        // msg.append("    -file    <file>              ''" + lSep);
-        // msg.append("    -f       <file>              ''" + lSep);
-//        msg.append("  --<property>=<value>         use value for given property\n");
-//        msg.append("  --propertyfile=<name>        load all properties from file\n");
-        // msg.append("  -keep-going, -k        execute all targets that do not depend"
-        // + lSep);
-        // msg.append("                         on failed target(s)" + lSep);
-        // msg.append("  -inputhandler <class>  the class which will handle input requests"
-        // + lSep);
-        // msg.append("  -nice  number          A niceness value for the main thread:"
-        // + lSep
-        // +
-        // "                         1 (lowest) to 10 (highest); 5 is the default"
-        // + lSep);
-        // msg.append("  -nouserlib             Run ant without using the jar files from"
-        // + lSep
-        // + "                         ${user.home}/.ant/lib" + lSep);
-        // msg.append("  -noclasspath           Run ant without using CLASSPATH"
-        // + lSep);
-        // msg.append("  -autoproxy             Java1.5+: use the OS proxy settings"
-        // + lSep);
-        // msg.append("  -main <class>          override Ant's normal entry point");
+        msg.append("  -h, --help                   print this message\n");
+        msg.append("  -v, --verbose                verbose logging\n");
         System.out.println(msg.toString());
     }
